+++ packages/lib/src/model/Model.ts
@@ -17,13 +17,7 @@ import {
   modelInitializersSymbol,
   modelPropertiesSymbol,
 } from "./modelSymbols"
-import {
-  ModelProps,
-  ModelPropsToCreationData,
-  ModelPropsToData,
-  noTypeChecker,
-  OptionalModelProps,
-} from "./prop"
+import { ModelProps, ModelPropsToCreationData, ModelPropsToData, OptionalModelProps } from "./prop"
 import { assertIsModelClass } from "./utils"
 
 declare const propsDataSymbol: unique symbol
@@ -156,12 +150,12 @@ function internalModel<TProps extends ModelProps, TBaseModel extends AnyModel>(
 
   // create type checker if needed
   let dataTypeChecker: LateTypeChecker | undefined
-  if (Object.values(composedModelProps).some(mp => mp.typeChecker !== noTypeChecker)) {
+  if (Object.values(composedModelProps).some(mp => !!mp.typeChecker)) {
     const typeCheckerObj: {
       [k: string]: any
     } = {}
     for (const [k, mp] of Object.entries(composedModelProps)) {
-      typeCheckerObj[k] = mp.typeChecker === noTypeChecker ? typesUnchecked() : mp.typeChecker
+      typeCheckerObj[k] = !mp.typeChecker ? typesUnchecked() : mp.typeChecker
     }
     dataTypeChecker = typesObject(() => typeCheckerObj) as any
   }