+++ packages/lib/src/typeChecking/array.ts
@@ -1,7 +1,7 @@
 import { isArray } from "../utils"
-import { resolveTypeChecker } from "./resolveTypeChecker"
-import { AnyType, ArrayType } from "./schemas"
-import { lateTypeChecker, TypeChecker } from "./TypeChecker"
+import { resolveStandardType, resolveTypeChecker } from "./resolveTypeChecker"
+import { AnyStandardType, AnyType, ArrayType } from "./schemas"
+import { getTypeInfo, lateTypeChecker, TypeChecker, TypeInfo, TypeInfoGen } from "./TypeChecker"
 import { TypeCheckError } from "./TypeCheckError"
 
 /**
@@ -17,29 +17,48 @@ import { TypeCheckError } from "./TypeCheckError"
  * @returns
  */
 export function typesArray<T extends AnyType>(itemType: T): ArrayType<T> {
+  const typeInfoGen: TypeInfoGen = t => new ArrayTypeInfo(t, resolveStandardType(itemType))
+
   return lateTypeChecker(() => {
     const itemChecker = resolveTypeChecker(itemType)
 
     const getTypeName = (...recursiveTypeCheckers: TypeChecker[]) =>
       `Array<${itemChecker.getTypeName(...recursiveTypeCheckers, itemChecker)}>`
 
-    const thisTc: TypeChecker = new TypeChecker((array, path) => {
-      if (!isArray(array)) {
-        return new TypeCheckError(path, getTypeName(thisTc), array)
-      }
+    const thisTc: TypeChecker = new TypeChecker(
+      (array, path) => {
+        if (!isArray(array)) {
+          return new TypeCheckError(path, getTypeName(thisTc), array)
+        }
 
-      if (!itemChecker.unchecked) {
-        for (let i = 0; i < array.length; i++) {
-          const itemError = itemChecker.check(array[i], [...path, i])
-          if (itemError) {
-            return itemError
+        if (!itemChecker.unchecked) {
+          for (let i = 0; i < array.length; i++) {
+            const itemError = itemChecker.check(array[i], [...path, i])
+            if (itemError) {
+              return itemError
+            }
           }
         }
-      }
 
-      return null
-    }, getTypeName)
+        return null
+      },
+      getTypeName,
+      typeInfoGen
+    )
 
     return thisTc
-  }) as any
+  }, typeInfoGen) as any
+}
+
+/**
+ * `types.array` type info.
+ */
+export class ArrayTypeInfo extends TypeInfo {
+  get itemTypeInfo(): TypeInfo {
+    return getTypeInfo(this.itemType)
+  }
+
+  constructor(thisType: AnyStandardType, readonly itemType: AnyStandardType) {
+    super(thisType)
+  }
 }